Styles applied to body were not rendering on inital load
There is an error in the console about mismatch classnames which
would seem to be about issues with SSR styled-components. But the
setup seems legit. I wondered if it was because its bad to put
styled-components within _document.js. And removing fixes the FOUC
but the console error persists. So they must be unrelated

Also tidied some other bits up