Setting posts with state not props
We were getting the posts on getInitialProps which happened server side.
But there is little advantage of doing this as most of the posts contain
stuff that only loads client side anyway (e.g. Instagram embeds).

The other problem is to load the page faster we want to be able to load
the posts a few at a time (paginated) so using state means we can
progressively load more posts and by using state can progressivley
append more easiy with react component lifecycles.

Also the next thing to build in after this is using Pusher so that the
client can listen for new posts and alert the user when a new post has
been published.