Fixed sticky Key Moments sidebar
Finally managed to get this working.

The problem is to use position:sticky; it has to be in an element that
has the same height as the body. But because we're using floats this is
hard. The JS I had written was flakey mainly because the leaderboard
takes longer to load and effects the height.

So new tack is to make the parent have the same height as the body but
this changes as elements are loaded into the page lazily (instagram
embeds, new posts etc) so I am checking the height everytime the user
scrolls which I guess is maybe overkill but it works. I had to subtract
a bit off otherwise the height grows constantly which is bad.