Get all posts show 5
So we can have all the key events, loading all posts (well first 100) with getInitialProps but then only showing the first 5 and showing more on click.

The key posts links arent quite right yet, need to pass it the array of displayed post so it can work out what is present and what its not. So can do page anchors or page links respectively